chore: Add login and logout functionality
The code changes in this commit add the login and logout functionality to the user module. It includes the implementation of the login and logout routes in the `routes.ts` file, as well as the corresponding controller functions in the `controller.ts` file. This allows users to authenticate and generate access tokens for subsequent requests.